moved 'organizations associated with DMP' into 'Researchers' section
