+++ oar-dmp/src/app/dmp-form/dmp-form.component.html
@@ -5,7 +5,7 @@
   <div class="dmpSection" >
     <table>
       <tr>
-        <td><label for="name">Record Name: </label></td>
+        <td><label for="name">DMP Short Name: </label></td>
         <td><div [ngClass]="nameClass"><input id="name" type="text" [formControl]="name" [readOnly]="nameDisabled"></div></td>
       </tr>
     </table>
