+++ oar-dmp/src/app/form-components/basic-info/basic-info.component.html
@@ -88,95 +88,8 @@
       <td><textarea name="projectDescription" cols="80" rows="5" placeholder="Description behind the science of the project"
         [formControl]="basicInfoForm.controls['projectDescription']"></textarea></td>
     </tr>
-    <tr>
-      <td>Select an organization associated with this DMP</td>
-      <td>
-        <mat-form-field class="example-full-width" appearance="fill" style="width:400px;">
-          <mat-label>NIST Organization</mat-label>
-          <input type="text"
-                 placeholder="Pick one"
-                 aria-label="NIST Organization"
-                 matInput
-                 [formControl]="basicInfoForm.controls['nistOrganization']"
-                 [matAutocomplete]="nistOrganization"
-                 >
-          <mat-autocomplete 
-            #nistOrganization="matAutocomplete"
-            [displayWith]="displaySelectedOrganization"
-          >
-            <mat-option *ngFor="let option of fltr_NIST_Org | async" [value]="option">
-              {{option.name}}
-            </mat-option>
-          </mat-autocomplete>
-        </mat-form-field>
-      </td>
-    </tr>
+    
     
   </table>
-  <div class="dmpSubSection">
-    <div class="sub-category">Organization(s) Associated With This DMP</div>  
-    <p class="instruction"> Add / Remove Organization(s) associated with this DMP</p>
-    
-    <article class="table-header">
-      
-      <!-- 
-        set button type to 'button' to prevent it from acting like a submit button and triggering
-        updateDMP in dmp.services.ts 
-      -->
-      <button type="button" class="button-remove-rows" mat-button (click)="removeSelectedRows()" [disabled]="disableRemove">Remove Selected</button>
-      <button type="button" class="button-clear-table" mat-button (click)="clearTable()" [disabled]="disableClear">Clear</button>
-      <button type="button" class="button-add-row" mat-button (click)="addRow()" [disabled]="disableAdd">Add</button>
-    </article>
-    <p *ngIf="errorMessage" class="error-msg">{{ errorMessage }}</p>
-    
   
-    <table mat-table [dataSource]="dmpOrganizations" class="mat-elevation-z8 form-table" id="dmpOrganizations">
-    
-      <!-- 
-        iterate over columns 
-      -->
-      <ng-container [matColumnDef]="col.key" *ngFor="let col of columnsSchema">
-        <!-- 
-          table headers 
-        -->
-        <th mat-header-cell *matHeaderCellDef>  {{ col.label }} </th>
-        <!-- 
-          table cells 
-        -->
-        <td mat-cell *matCellDef="let element">  
-          <!-- table cell display default cell display -->
-          <div [ngSwitch]="col.type" *ngIf="!element.isEdit">
-            <!-- Switch if the column isSelected to display the checkbox in the table cell -->
-            <ng-container *ngSwitchCase="'isSelected'">
-              <mat-checkbox
-                (change)="element.isSelected = $event.checked"
-              ></mat-checkbox>
-            </ng-container>
-  
-            <!-- Switch if the column isEdit to display the Edit and delete buttons in the table cell -->
-            <div class="btn-edit" *ngSwitchCase="'isEdit'">
-              <button
-                mat-button
-                class="button-remove"
-                (click)="removeRow(element.id)"
-              >
-                Delete
-              </button>
-            </div>
-            
-            <!-- default cell -->
-            <span *ngSwitchDefault>
-              {{ element[col.key] }}
-            </span>
-          </div>
-  
-         
-  
-        </td>
-      </ng-container>
-    
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-    </table>
-  </div>
 </div>
