+++ oar-dmp/src/app/form-components/keywords/keywords.component.html
@@ -83,7 +83,7 @@
     <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
     <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
   </table>
-
+  <br/>
