+++ oar-dmp/src/app/form-components/personel/personel.component.ts
@@ -9,6 +9,8 @@ import { NistContact } from '../../types/nist-contact'
 import { Validators, FormBuilder } from '@angular/forms';
 import { defer, map, of, startWith } from 'rxjs';
 import { DMP_Meta } from '../../types/DMP.types';
+import { ORGANIZATIONS } from '../../types/mock-organizations';
+import { NistOrganization } from 'src/app/types/nist-organization';
 
 import {Observable} from 'rxjs';
 
@@ -68,12 +70,60 @@ const COLUMNS_SCHEMA = [
   },
 ]
 
+export interface dmpOgranizations {
+  org_id:number;
+  dmp_organization: string;  
+  id: number;
+  isEdit: boolean;
+}
+
+const ORG_COL_SCHEMA = [
+  {
+    key: 'isSelected',
+    type: 'isSelected',
+    label: '',
+  },
+  {
+    key: 'org_id',
+    type: 'text',
+    label: 'Org ID',
+  },
+  {
+    key: 'dmp_organization',
+    type: 'text',
+    label: 'Organization(s)',
+  },
+  // Edit button column
+  {
+    key: 'isEdit',
+    type: 'isEdit',
+    label: '',
+  },
+]
+
 @Component({
   selector: 'app-personel',
   templateUrl: './personel.component.html',
   styleUrls: ['./personel.component.scss', '../form-table.scss']
 })
 export class PersonelComponent implements OnInit {
+  // ================================
+  // used for organizations table
+  // ================================
+  org_disableAdd:boolean = true;
+  org_disableClear:boolean = true;
+  org_disableRemove:boolean = true;
+  org_errorMessage: string = '';
+  dmpOrganizations: dmpOgranizations[] = []
+  org_displayedColumns: string[] = ORG_COL_SCHEMA.map((col) => col.key);
+  org_columnsSchema: any = ORG_COL_SCHEMA;
+  fltr_NIST_Org!: Observable<NistOrganization[]>;
+  //List of all nist organizations from NIST directory
+  nistOrganizations: any = null;
+  crntOrgID:number = 0;
+  crntOrgName:string = "";
+  
+  // ================================  
 
   disableAdd:boolean = true;
   disableClear:boolean = true;
@@ -138,7 +188,9 @@ export class PersonelComponent implements OnInit {
       dmp_contributor:            [''],
       nistContactFirstName:       [''],
       nistContactLastName:        [''],      
-      contributors:               [[]]
+      contributors:               [[]],
+      nistOrganization: [],
+      organizations: [[]]
     }
   );
 
@@ -149,6 +201,16 @@ export class PersonelComponent implements OnInit {
   // the form. Here you could do any data transformation you need.
   @Input()
   set initialDMP_Meta(personel: DMP_Meta) {
+    // loop over organizations array sent from the server and populate local copy of 
+    // organizations aray in order to populate the table of organizations in the GUI interface
+    personel.organizations.forEach( 
+      (org, index) => {        
+        this.dmpOrganizations.push({id:index, org_id:org.ORG_ID, dmp_organization:org.name, isEdit:false});
+        this.disableClear=false;
+        this.disableRemove=false;
+      }
+      
+    );
     // loop over resources array sent from the server and populate local copy of 
     // resources array to populate the table of resources in the user interface
     personel.contributors.forEach(
@@ -206,7 +268,8 @@ export class PersonelComponent implements OnInit {
                                       lastName: formValue.nistContactLastName,
                                       orcid:formValue.primNistContactOrcid
                                     },
-            contributors:           formValue.contributors
+            contributors:           formValue.contributors,
+            organizations:formValue.organizations
 
           }
         )
@@ -220,6 +283,11 @@ export class PersonelComponent implements OnInit {
      * NOTE Comment below when woking with API
      */
     this.getNistContacts();
+
+    /**
+     * NOTE Comment below when woking with API
+     */
+    this.getNistOrganizations();
 
   }  
 
@@ -681,7 +749,7 @@ export class PersonelComponent implements OnInit {
     }
 
     this.errorMessage = '';
-    this.resetTable();
+    this.org_resetTable();
     this.dmpContributors.forEach((element)=>{
         if(element.id === e.id){
           element.isEdit = false;
@@ -725,7 +793,173 @@ export class PersonelComponent implements OnInit {
       primNistContactOrcid:       ""
     });
     this.clearTable();
+    this.org_clearTable();
   }
+
+  // ==================================================
+  // ==================================================
+  // ==================================================
+
+   /**
+   * This function gets hard coded NIST organizations
+   * Used when not working with an API for NIST people service database
+   */
+   getNistOrganizations(){    
+    // this.getNistOrganizationsFromAPI();
+    this.getNistOrganizationsNoAPI();
+
+  }
+
+  getNistOrganizationsNoAPI(){
+    //ORGANIZATIONS is declared in '../../types/mock-organizations'
+    this.nistOrganizations = ORGANIZATIONS;
+    this.fltr_NIST_Org = this.personelForm.controls['nistOrganization'].valueChanges.pipe(
+      startWith(''),
+      map(anOrganization => {
+        const orgName = typeof anOrganization ==='string' ? anOrganization : anOrganization?.name
+        var res = orgName ? this._filter_org(orgName as string):this.nistOrganizations.slice();
+        if (res.length ===1){
+          this.crntOrgID = anOrganization.ORG_ID;
+          this.crntOrgName = anOrganization.name;
+
+          this.org_disableAdd = false;
+        }
+        else{
+          this.org_disableAdd = true;
+        }
+        return res;
+      }
+
+      )
+
+    );
+
+  }
+
+  displaySelectedOrganization(org:NistOrganization):string{
+    var res = org && org.ORG_ID? org.name : '';
+    return res;
+
+  }
+
+  org_removeSelectedRows() {
+    //assign unselected rows to dmpOrganizations
+    this.dmpOrganizations = this.dmpOrganizations.filter((u: any) => !u.isSelected);
+    //reset the table 
+    this.org_resetTable();
+    //repopulate the table with what's left in the array
+    this.rePopulateOrgs();
+
+    if (this.dmpOrganizations.length === 0){
+      // If the table is empty disable clear and remove buttons
+      this.org_disableClear=true;
+      this.org_disableRemove=true;
+    }
+  }
+
+  org_resetTable(){
+    this.personelForm.patchValue({
+      organizations:[]
+    })
+  }
+
+  org_clearTable(){
+    this.dmpOrganizations = []
+    this.org_resetTable();
+    this.org_disableAdd=true;
+    this.org_disableClear=true;
+    this.org_disableRemove=true;
+  }
+
+  org_addRow(){
+    const newRow = {
+      id: Date.now(),      
+      org_id:this.crntOrgID,
+      dmp_organization: this.crntOrgName,
+      isEdit: false,
+    };
+    // check that if any org id or org name is undefined 
+    // - this can happen if user types in the search box but does not select 
+    //    an actual organization from the drop down menu
+
+    if (typeof newRow.org_id === "undefined" || typeof newRow.dmp_organization === "undefined"){
+      this.org_errorMessage = "Select an existing NIST Organization";
+      return;
+
+    }
+
+    // Check if selected organization is already in the table
+    var selRow = this.dmpOrganizations.filter((u) => u.org_id === newRow.org_id);
+    if (selRow.length > 0){
+      this.org_errorMessage = "The selected Organization is already associated with this DMP.";
+      return;
+    }
+    //add new row to the dmpOrganizations array
+    this.dmpOrganizations = [newRow, ...this.dmpOrganizations]
+
+    //reset the table
+    this.org_resetTable();
+
+    // re-populate the table with entries from dmpOrganizations array 
+    // and update the form metadata
+    this.rePopulateOrgs();
+
+    this.org_disableAdd=true;
+    this.org_disableClear=false;
+    this.org_disableRemove=false;
+  }
+
+  private rePopulateOrgs(){
+    this.dmpOrganizations.forEach(
+      (org)=>{
+        this.personelForm.value['organizations'].push(
+          {
+            ORG_ID:org.org_id,
+            name:org.dmp_organization
+          }
+        )
+      }
+    )
+  }
+
+  org_removeRow(id:any) {
+    var selRow = this.dmpOrganizations.filter((u) => u.id === id); 
+
+    // update the form metadata
+    this.personelForm.value['organizations'].forEach(
+      (value:NistOrganization, index:number)=>{
+        selRow.forEach(
+          (org)=>{
+            if (value.ORG_ID === org.org_id){
+              //remove selected organization
+              this.personelForm.value['organizations'].splice(index,1);
+            }
+          }
+        )
+      }
+    )
+
+    // remove from the display table
+    this.dmpOrganizations = this.dmpOrganizations.filter((u) => u.id !== id);
+  }
+
+  private _filter_org(nistOrg:string): NistOrganization[] {
+    // add button should be disabled while filtering is being performed
+    // and should be enabled only when an existing organization has been selected
+    this.org_disableAdd = true;
+
+    const filterValues = nistOrg.toLowerCase()
+    var searchRes;
+    searchRes = this.nistOrganizations.filter(
+      (option:any) => option.name.toLowerCase().includes(filterValues)
+    );
+
+    return searchRes;
+
+  }
+
+
+
