+++ oar-dmp/src/app/form-components/technical-requirements/technical-requirements.component.html
@@ -175,5 +175,6 @@
   
     <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
     <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-  </table>    
+  </table>
+  <br/>    
 </div>
